Global function estimate the recombination map

Global function estimate the recombination map

recombination_map(
  x,
  method = "loess",
  K = 5,
  boot = 1000,
  n_cores = 1,
  smoothing = numeric(),
  n_resampling = 1000,
  calibration_range = c(0.2, 0.5),
  degree = 2,
  windows = 10^5,
  set_negative_values = 0,
  verbose = TRUE
)

Arguments

x

a `marey_map` object.

method

an interpolation method, either 'loess' or 'splines' (default = 'loess').

K

number of clusters to subset in K-fold cross-validation.

boot

number of bootstraps to estimate the confidence interval.

n_cores

number of cores to parallelize.

smoothing

(optional) smoothing parameter, if you do not want to calibrate it.

n_resampling

number of iterations in the cross-validation procedure.

calibration_range

range of the smoothing parameter space to explore.

degree

(optional) the degree parameter of the polynomial used in the 'loess' method.

windows

size or coordinates of the windows along the chromosome: either integer in basepairs (default = 10^5) or a data.frame of start/end coordinates.

set_negative_values

the value (e.g. 0 or NA) to use for negative values in recombination rates (default = 0).

verbose

Whether to print messages and progress bars at each step (default = TRUE).

Value

a `marey_map` object with the estimated recombination map.